Require Western Arabic numerals for max_rows
Previously, any Unicode character with the General_Category property of
Decimal_Number was permitted as the value of max_rows.  This included Eastern
Arabic numerals like '٤٠' and Devanagari numerals like '४०' (both represent the
Western Arabic '40').  However, max_rows was later being used with the numeric
equality operator (==), which only understands Western Arabic numerals.  This
was fixed by replacing \d in the validation regex with [0-9] so that only '0'
through '9' are valid and no other of the Decimal_Number characters.